moved `StringWarning` to `pc`
