+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/sourcecode/SourceCodeState.scala
@@ -19,7 +19,7 @@ package org.alephium.ralph.lsp.pc.sourcecode
 import org.alephium.ralph.lsp.access.compiler.RalphParserExtension
 import org.alephium.ralph.lsp.access.compiler.ast.Tree
 import org.alephium.ralph.lsp.access.compiler.message.CompilerMessage
-import org.alephium.ralph.lsp.access.compiler.message.warning.StringWarning
+import org.alephium.ralph.lsp.pc.sourcecode.warning.StringWarning
 import org.alephium.ralph.lsp.pc.util.URIUtil
 import org.alephium.ralph.{CompiledScript, CompiledContract}
