+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/sourcecode/warning/StringWarning.scala
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see http://www.gnu.org/licenses/.
 
-package org.alephium.ralph.lsp.access.compiler.message.warning
+package org.alephium.ralph.lsp.pc.sourcecode.warning
 
 import org.alephium.ralph
 import org.alephium.ralph.SourceIndex
